Project structure converted to package-mode.
