+++ yapam/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+"""Yet another ammo generator for Yandex.Tank (c).
+
+for additional info see README.md
+"""
+
+__version__ = '0.1.0'
+__all__ = [
+    'armory', 'config', 'phantom'
+]
+
+__author__ = 'Aleksey Devyatkin <devyatkin.av@ya.ru>'
